requirement.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-mistralai
langchain-nvidia-ai-endpoints